+++ src/rust/lib/debug-scenes/Cargo.toml
@@ -21,6 +21,8 @@ wasm-bindgen         = { version = "=0.2.58" , features = ["nightly"] }
 nalgebra             = { version = "0.19.0"                           }
 js-sys               = { version = "0.3.28"                           }
 
+im-rc = "14.3.0"
+
 [dependencies.web-sys]
 version = "0.3.4"
 features = [
