+++ src/rust/lib/debug-scenes/src/shapes.rs
@@ -16,6 +16,7 @@ use ensogl::display::shape::primitive::system::ShapeSystem;
 use ensogl::display::shape::Var;
 use ensogl::display::world::*;
 use ensogl::system::web;
+use graph::node;
 use graph::node::Node;
 use graph::node::NodeRegistry;
 use nalgebra::Vector2;
@@ -29,7 +30,7 @@ use ensogl::control::io::mouse;
 use enso_frp::core::node::class::EventEmitterPoly;
 use ensogl_system_web::StyleSetter;
 use ensogl::display::layout::alignment;
-
+use wasm_bindgen::JsCast;
 
 
 #[derive(Debug,Clone)]
@@ -82,91 +83,7 @@ pub fn run_example_shapes() {
     init(&World::new(&web::get_html_element_by_id("root").unwrap()));
 }
 
-pub mod icons {
-    use super::*;
 
-    pub fn history() -> AnyShape {
-        let radius_diff    = 0.5.px();
-        let corners_radius = 2.0.px();
-        let width_diff     = &corners_radius * 3.0;
-        let offset         = 2.px();
-        let width          = 32.px();
-        let height         = 16.px();
-        let persp_diff1    = 6.px();
-
-        let width2          = &width  - &width_diff;
-        let width3          = &width2 - &width_diff;
-        let corners_radius2 = &corners_radius  - &radius_diff;
-        let corners_radius3 = &corners_radius2 - &radius_diff;
-        let persp_diff2     = &persp_diff1 * 2.0;
-
-        let rect1 = Rect((&width ,&height)).corners_radius(&corners_radius);
-        let rect2 = Rect((&width2,&height)).corners_radius(&corners_radius2).translate_y(&persp_diff1);
-        let rect3 = Rect((&width3,&height)).corners_radius(&corners_radius3).translate_y(&persp_diff2);
-
-        let rect3 = rect3 - rect2.translate_y(&offset);
-        let rect2 = rect2 - rect1.translate_y(&offset);
-
-        let rect1 = rect1.fill(Srgba::new(0.26, 0.69, 0.99, 1.00));
-        let rect2 = rect2.fill(Srgba::new(0.26, 0.69, 0.99, 0.6));
-        let rect3 = rect3.fill(Srgba::new(0.26, 0.69, 0.99, 0.4));
-
-        let icon = (rect3 + rect2 + rect1).translate_y(-persp_diff2/2.0);
-        icon.into()
-    }
-}
-
-fn ring_angle<R,W,A>(inner_radius:R, width:W, angle:A) -> AnyShape
-where R : Into<Var<Distance<Pixels>>>,
-      W : Into<Var<Distance<Pixels>>>,
-      A : Into<Var<Angle<Radians>>> {
-    let inner_radius = inner_radius.into();
-    let width        = width.into();
-    let angle        = angle.into();
-
-    let angle2  = &angle / 2.0;
-    let radius  = &width / 2.0;
-    let inner   = Circle(&inner_radius);
-    let outer   = Circle(&inner_radius + &width);
-    let section = Plane().cut_angle(&angle);
-    let corner1 = Circle(&radius).translate_y(inner_radius + radius);
-    let corner2 = corner1.rotate(&angle2);
-    let corner1 = corner1.rotate(-&angle2);
-    let ring    = &outer - &inner;
-    let pie     = &ring * &section;
-    let out     = &pie + &corner1 + &corner2;
-    let out     = out.fill(Srgba::new(0.9,0.9,0.9,1.0));
-    out.into()
-}
-
-fn nodes_shape() -> AnyShape {
-    let node_radius = 32.0;
-    let border_size = 16.0;
-
-    let node = Circle(node_radius.px());
-    let node = node.fill(Srgb::new(0.97,0.96,0.95));
-    let bg   = Circle((node_radius*2.0).px());
-    let bg   = bg.fill(Srgb::new(0.91,0.91,0.90));
-
-    let shadow2 = Circle((node_radius + border_size).px());
-    let shadow2_color = LinearGradient::new()
-        .add(0.0,Srgba::new(0.0,0.0,0.0,0.0).into_linear())
-        .add(1.0,Srgba::new(0.0,0.0,0.0,0.14).into_linear());
-    let shadow2_color = SdfSampler::new(shadow2_color).max_distance(border_size).slope(Slope::Exponent(4.0));
-    let shadow2       = shadow2.fill(shadow2_color);
-
-    let selection = Circle((node_radius - 1.0).px() + border_size.px() * "input_selection");
-    let selection = selection.fill(Srgba::new(0.22,0.83,0.54,1.0));
-
-    let loader_angle : Var<Angle<Radians>> = "Radians(clamp(input_time/2000.0 - 1.0) * 1.99 * PI)".into();
-    let loader_angle2 = &loader_angle / 2.0;
-    let loader        = ring_angle((node_radius).px(), (border_size).px(), loader_angle);
-    let loader        = loader.rotate(loader_angle2);
-    let loader        = loader.rotate("Radians(input_time/200.0)");
-    let icon          = icons::history();
-    let out           = loader + selection + shadow2 + node + icon;
-    out.into()
-}
 
 fn mouse_pointer() -> AnyShape {
     let radius  = 10.px();
@@ -187,6 +104,13 @@ use ensogl::control::event_loop::AnimationLoop;
 use ensogl::control::event_loop::TimeInfo;
 use ensogl::control::event_loop::FixedFrameRateSampler;
 use ensogl::animation::physics::inertia::DynInertiaSimulator;
+use ensogl::data::OptVec;
+use im_rc as im;
+
+#[derive(Debug,Default,Clone)]
+pub struct NodeSet {
+    vec : Rc<RefCell<OptVec<Node>>>
+}
 
 
 fn init(world: &World) {
@@ -196,12 +120,14 @@ fn init(world: &World) {
     let navigator = Navigator::new(&scene,&camera);
 
 
-    let node_shape_system             = ShapeSystem::new(world,&nodes_shape());
+    let node_shape_system             = ShapeSystem::new(world,&node::shape());
+    let node_selection_buffer         = node_shape_system.add_input("selection" , 0.0);
+
+
     let pointer_shape_system          = ShapeSystem::new(world,&mouse_pointer());
     let pointer_position_buffer       = pointer_shape_system.add_input("position" , Vector2::<f32>::new(0.0,0.0));
     let pointer_selection_size_buffer = pointer_shape_system.add_input("selection_size" , Vector2::<f32>::new(0.0,0.0));
 
-    let node_selection_buffer         = node_shape_system.add_input("selection" , 0.0);
 
 
     let shape = scene.dom.shape().current();
@@ -245,23 +171,23 @@ fn init(world: &World) {
 
 
     world.add_child(&node1);
-    world.add_child(&node2);
-    world.add_child(&node3);
+//    world.add_child(&node2);
+//    world.add_child(&node3);
 
     node1.mod_position(|t| {
         t.x += 200.0;
         t.y += 200.0;
     });
 
-    node2.mod_position(|t| {
-        t.x += 300.0;
-        t.y += 300.0;
-    });
-
-    node3.mod_position(|t| {
-        t.x += 400.0;
-        t.y += 200.0;
-    });
+//    node2.mod_position(|t| {
+//        t.x += 300.0;
+//        t.y += 300.0;
+//    });
+//
+//    node3.mod_position(|t| {
+//        t.x += 400.0;
+//        t.y += 200.0;
+//    });
 
 
     let nodes = vec![node1,node2,node3];
@@ -297,10 +223,30 @@ fn init(world: &World) {
         node1_selection         = source::<f32>           ();
         node1_selection_a       = source::<f32>           ();
 
+        nodes = source::<NodeSet> ();
+        add_node = source::<()> ();
+        new_node = add_node.map2(&mouse.position, enclose!((world,node_registry) move |_,pos| {
+            let node = Node::new(&node_registry);
+            world.add_child(&node);
+            node.mod_position(|t| {
+                t.x += pos.x as f32;
+                t.y += pos.y as f32;
+            });
+            Some(node)
+        }));
+
+        nodes_update = nodes.map2(&new_node, |node_set,new_node| {
+            new_node.for_each_ref(|node| {
+                node_set.vec.borrow_mut().insert(node.clone_ref());
+            })
+        });
+
         debug = node1_selection_a.map(|t| {println!("{:?}",t);})
 
 
     }
+
+//    nodes_update.event.display_graphviz();
 
     mouse.position.map("pointer_position", enclose!((pointer) move |p| {
         pointer.sprite.get().for_each(|sprite| {
@@ -330,23 +276,26 @@ fn init(world: &World) {
         match target {
             display::scene::Target::Background => {}
             display::scene::Target::Symbol {symbol_id, instance_id} => {
-//                node_selection_buffer.at((*instance_id as usize).into()).set(1.0);
                 let br   = node_registry.map.borrow();
                 let node = br.get(&(*instance_id as usize));
                 node.for_each(|node| {
                     node.selection.event.emit(());
                 })
-//                node_shape_system.get(*instance_id as usize).for_each(|sprite|{
-//                    let selection = node_selection_buffer.at(sprite.instance_id());
-//                    selection.set(1.0);
-//                })
             }
         }
         println!("SELECTING {:?}", target);
     }));
 
 
-
+    let c: Closure<dyn Fn(JsValue)> = Closure::wrap(Box::new(move |val| {
+        let val = val.unchecked_into::<web_sys::KeyboardEvent>();
+        let key = val.key();
+        if      key == "n" {
+            add_node.event.emit(());
+        }
+    }));
+    web::document().add_event_listener_with_callback("keydown",c.as_ref().unchecked_ref()).unwrap();
+    c.forget();
